Refactor database.jf file
Moved the queries for users, properties and reservations to their own specific files. Refacoted routes.js files to reflect these chanfes.
